+++ rocketpy/motors/hybrid_motor.py
@@ -1,4 +1,4 @@
-from ..mathutils.function import funcify_method, reset_funcified_methods
+from ..mathutils.function import Function, funcify_method, reset_funcified_methods
 from ..plots.hybrid_motor_plots import _HybridMotorPlots
 from ..prints.hybrid_motor_prints import _HybridMotorPrints
 from .liquid_motor import LiquidMotor
@@ -214,10 +214,10 @@ class HybridMotor(Motor):
             also be given as a callable function, whose argument is time in
             seconds and returns the thrust supplied by the motor in the
             instant. If a string is given, it must point to a .csv or .eng file.
-            The .csv file shall contain no headers and the first column must
-            specify time in seconds, while the second column specifies thrust.
-            Arrays may also be specified, following rules set by the class
-            Function. Thrust units are Newtons.
+            The .csv file can contain a single line header and the first column
+            must specify time in seconds, while the second column specifies
+            thrust. Arrays may also be specified, following rules set by the
+            class Function. Thrust units are Newtons.
 
             .. seealso:: :doc:`Thrust Source Details </user/motors/thrust>`
         dry_mass : int, float
@@ -372,7 +372,9 @@ class HybridMotor(Motor):
         self.exhaust_velocity : Function
             Gas exhaust velocity of the motor.
         """
-        return self.total_impulse / self.propellant_initial_mass
+        return Function(
+            self.total_impulse / self.propellant_initial_mass
+        ).set_discrete_based_on_model(self.thrust)
 
     @funcify_method("Time (s)", "Mass (kg)")
     def propellant_mass(self):