+++ rocketpy/simulation/__init__.py
@@ -1,2 +1,3 @@
 from .dispersion import Dispersion
 from .flight import Flight
+from .flight_data_importer import FlightDataImporter