+++ tests/test_function.py
@@ -73,8 +73,8 @@ def test_getters(func_from_csv, func_2d_from_csv):
     assert func_2d_from_csv.get_outputs() == ["Scalar"]
     assert func_2d_from_csv.get_interpolation_method() == "shepard"
     assert func_2d_from_csv.get_extrapolation_method() == "natural"
-    assert np.isclose(func_2d_from_csv.get_value(0, 0), 0.0, atol=1e-6)
-    assert np.isclose(func_2d_from_csv.get_value_opt(0, 0), 0.0, atol=1e-6)
+    assert np.isclose(func_2d_from_csv.get_value(0.1, 0.8), 0.058, atol=1e-6)
+    assert np.isclose(func_2d_from_csv.get_value_opt(0.1, 0.8), 0.058, atol=1e-6)
 
 
 def test_setters(func_from_csv, func_2d_from_csv):
@@ -189,127 +189,6 @@ def test_extrapolation_methods(linear_func):
     assert np.isclose(linear_func.get_value(-1), -1, atol=1e-6)
 
 
-@pytest.mark.parametrize("a", [-1, 0, 0.5, 1, 2, 2.5, 3.5, 4, 5])
-@pytest.mark.parametrize("b", [-1, 0, 0.5, 1, 2, 2.5, 3.5, 4, 5])
-def test_integral_linear_interpolation(linearly_interpolated_func, a, b):
-    """Test the integral method of the Function class.
-
-    Parameters
-    ----------
-    linear_func : rocketpy.Function
-        A Function object created from a list of values.
-    """
-    # Test integral
-    assert isinstance(linearly_interpolated_func.integral(a, b, numerical=True), float)
-    assert np.isclose(
-        linearly_interpolated_func.integral(a, b, numerical=False),
-        linearly_interpolated_func.integral(a, b, numerical=True),
-        atol=1e-3,
-    )
-
-
-@pytest.mark.parametrize("func", ["linear_func", "spline_interpolated_func"])
-@pytest.mark.parametrize("a", [-1, -0.5, 0, 0.5, 1, 2, 2.5, 3.5, 4, 5])
-@pytest.mark.parametrize("b", [-1, -0.5, 0, 0.5, 1, 2, 2.5, 3.5, 4, 5])
-def test_integral_spline_interpolation(request, func, a, b):
-    """Test the integral method of the Function class.
-
-    Parameters
-    ----------
-    spline_func : rocketpy.Function
-        A Function object created from a list of values.
-    a : float
-        Lower limit of the integral.
-    b : float
-        Upper limit of the integral.
-    """
-    # Test integral
-    # Get the function from the fixture
-    func = request.getfixturevalue(func)
-    assert np.isclose(
-        func.integral(a, b, numerical=False),
-        func.integral(a, b, numerical=True),
-        atol=1e-3,
-    )
-
-
-@pytest.mark.parametrize(
-    "func_input, derivative_input, expected_derivative",
-    [
-        (1, 0, 0),  # Test case 1: Function(1)
-        (lambda x: x, 0, 1),  # Test case 2: Function(lambda x: x)
-        (lambda x: x**2, 1, 2),  # Test case 3: Function(lambda x: x**2)
-    ],
-)
-def test_differentiate(func_input, derivative_input, expected_derivative):
-    """Test the differentiate method of the Function class.
-
-    Parameters
-    ----------
-    func_input : function
-        A function object created from a list of values.
-    derivative_input : int
-        Point at which to differentiate.
-    expected_derivative : float
-        Expected value of the derivative.
-    """
-    func = Function(func_input)
-    assert isinstance(func.differentiate(derivative_input), float)
-    assert np.isclose(func.differentiate(derivative_input), expected_derivative)
-
-
-def test_get_value():
-    """Tests the get_value method of the Function class.
-    Both with respect to return instances and expected behaviour.
-    """
-    func = Function(lambda x: 2 * x)
-    assert isinstance(func.get_value(1), int or float)
-
-
-def test_identity_function():
-    """Tests the identity_function method of the Function class.
-    Both with respect to return instances and expected behaviour.
-    """
-
-    func = Function(lambda x: x**2)
-    assert isinstance(func.identity_function(), Function)
-
-
-def test_derivative_function():
-    """Tests the derivative_function method of the Function class.
-    Both with respect to return instances and expected behaviour.
-    """
-    square = Function(lambda x: x**2)
-    assert isinstance(square.derivative_function(), Function)
-
-
-def test_integral():
-    """Tests the integral method of the Function class.
-    Both with respect to return instances and expected behaviour.
-    """
-
-    zero_func = Function(0)
-    assert isinstance(zero_func.integral(2, 4, numerical=True), float)
-    assert zero_func.integral(2, 4, numerical=True) == 0
-
-    square = Function(lambda x: x**2)
-    assert isinstance
-    assert square.integral(2, 4, numerical=True) == -square.integral(
-        4, 2, numerical=True
-    )
-    assert square.integral(2, 4, numerical=False) == -square.integral(
-        4, 2, numerical=False
-    )
-
-
-def test_integral_function():
-    """Tests the integral_function method of the Function class.
-    Both with respect to return instances and expected behaviour.
-    """
-    zero_func = Function(0)
-    assert isinstance(zero_func, Function)
-
-
 @pytest.mark.parametrize("a", [-1, 0, 1])
 @pytest.mark.parametrize("b", [-1, 0, 1])
 def test_multivariable_dataset(a, b):
@@ -334,6 +213,55 @@ def test_multivariable_dataset(a, b):
 
     # Assert values
     assert np.isclose(func(a, b), a + b, atol=1e-6)
+
+
+@pytest.mark.parametrize(
+    "x,y,z_expected",
+    [
+        (1, 0, 0),
+        (0, 1, 0),
+        (0, 0, 1),
+        (0.5, 0.5, 1 / 3),
+        (0.25, 0.25, 25 / (25 + 2 * 5**0.5)),
+        ([0, 0.5], [0, 0.5], [1, 1 / 3]),
+    ],
+)
+def test_2d_shepard_interpolation(x, y, z_expected):
+    """Test the shepard interpolation method of the Function class."""
+    # Test plane x + y + z = 1
+    source = [(1, 0, 0), (0, 1, 0), (0, 0, 1)]
+    func = Function(
+        source=source, inputs=["x", "y"], outputs=["z"], interpolation="shepard"
+    )
+    z = func(x, y)
+    z_opt = func.get_value_opt(x, y)
+    assert np.isclose(z, z_opt, atol=1e-8).all()
+    assert np.isclose(z_expected, z, atol=1e-8).all()
+
+
+@pytest.mark.parametrize(
+    "x,y,z,w_expected",
+    [
+        (0, 0, 0, 1),
+        (1, 0, 0, 0),
+        (0, 1, 0, 0),
+        (0, 0, 1, 0),
+        (0.5, 0.5, 0.5, 1 / 4),
+        (0.25, 0.25, 0.25, 0.700632626832),
+        ([0, 0.5], [0, 0.5], [0, 0.5], [1, 1 / 4]),
+    ],
+)
+def test_3d_shepard_interpolation(x, y, z, w_expected):
+    """Test the shepard interpolation method of the Function class."""
+    # Test plane x + y + z + w = 1
+    source = [(1, 0, 0, 0), (0, 1, 0, 0), (0, 0, 1, 0), (0, 0, 0, 1)]
+    func = Function(
+        source=source, inputs=["x", "y", "z"], outputs=["w"], interpolation="shepard"
+    )
+    w = func(x, y, z)
+    w_opt = func.get_value_opt(x, y, z)
+    assert np.isclose(w, w_opt, atol=1e-8).all()
+    assert np.isclose(w_expected, w, atol=1e-8).all()
 
 
 @pytest.mark.parametrize("a", [-1, -0.5, 0, 0.5, 1])
@@ -378,6 +306,55 @@ def test_multivariable_function_plot(mock_show):
 
     # Assert plot
     assert func.plot() == None
+
+
+def test_set_discrete_2d():
+    """Tests the set_discrete method of the Function for
+    two dimensional domains.
+    """
+    func = Function(lambda x, y: x**2 + y**2)
+    discretized_func = func.set_discrete([-5, -7], [8, 10], [50, 100])
+
+    assert isinstance(discretized_func, Function)
+    assert isinstance(func, Function)
+    assert discretized_func.source.shape == (50 * 100, 3)
+    assert np.isclose(discretized_func.source[0, 0], -5)
+    assert np.isclose(discretized_func.source[0, 1], -7)
+    assert np.isclose(discretized_func.source[-1, 0], 8)
+    assert np.isclose(discretized_func.source[-1, 1], 10)
+
+
+def test_set_discrete_2d_simplified():
+    """Tests the set_discrete method of the Function for
+    two dimensional domains with simplified inputs.
+    """
+    source = [(1, 0, 0), (0, 1, 0), (0, 0, 1)]
+    func = Function(source=source, inputs=["x", "y"], outputs=["z"])
+    discretized_func = func.set_discrete(-1, 1, 10)
+
+    assert isinstance(discretized_func, Function)
+    assert isinstance(func, Function)
+    assert discretized_func.source.shape == (100, 3)
+    assert np.isclose(discretized_func.source[0, 0], -1)
+    assert np.isclose(discretized_func.source[0, 1], -1)
+    assert np.isclose(discretized_func.source[-1, 0], 1)
+    assert np.isclose(discretized_func.source[-1, 1], 1)
+
+
+def test_set_discrete_based_on_2d_model(func_2d_from_csv):
+    """Tests the set_discrete_based_on_model method with a 2d model
+    Function.
+    """
+    func = Function(lambda x, y: x**2 + y**2)
+    discretized_func = func.set_discrete_based_on_model(func_2d_from_csv)
+
+    assert isinstance(discretized_func, Function)
+    assert isinstance(func, Function)
+    assert np.array_equal(
+        discretized_func.source[:, :2], func_2d_from_csv.source[:, :2]
+    )
+    assert discretized_func.__interpolation__ == func_2d_from_csv.__interpolation__
+    assert discretized_func.__extrapolation__ == func_2d_from_csv.__extrapolation__
 
 
 @pytest.mark.parametrize(
@@ -478,3 +455,43 @@ def test_pow_arithmetic_priority(other):
     assert isinstance(other**func_lambda, Function)
     assert isinstance(func_array**other, Function)
     assert isinstance(other**func_array, Function)
+
+
+@pytest.mark.parametrize("alpha", [0.1, 0.5, 0.9])
+def test_low_pass_filter(alpha):
+    """Test the low_pass_filter method of the Function class.
+
+    Parameters
+    ----------
+    alpha : float
+        Attenuation coefficient, 0 < alpha < 1.
+    """
+    # Create a test function, sinus here
+    source = np.array(
+        [(1, np.sin(1)), (2, np.sin(2)), (3, np.sin(3)), (4, np.sin(4)), (5, np.sin(5))]
+    )
+    func = Function(source)
+
+    # Apply low pass filter
+    filtered_func = func.low_pass_filter(alpha)
+
+    # Check that the method works as intended and returns the right object with no issue
+    assert isinstance(filtered_func, Function), "The returned type is not a Function"
+    assert np.array_equal(
+        filtered_func.source[0], source[0]
+    ), "The initial value is not the expected value"
+    assert len(filtered_func.source) == len(
+        source
+    ), "The filtered Function and the Function have different lengths"
+    assert (
+        filtered_func.__interpolation__ == func.__interpolation__
+    ), "The interpolation method was unexpectedly changed"
+    assert (
+        filtered_func.__extrapolation__ == func.__extrapolation__
+    ), "The extrapolation method was unexpectedly changed"
+    for i in range(1, len(source)):
+        expected = alpha * source[i][1] + (1 - alpha) * filtered_func.source[i - 1][1]
+        assert np.isclose(filtered_func.source[i][1], expected, atol=1e-6), (
+            f"The filtered value at index {i} is not the expected value. "
+            f"Expected: {expected}, Actual: {filtered_func.source[i][1]}"
+        )